extract browse list types and transforms
